feat: 게임 프로그램 구조 변경(함수 -> 구조체) 및 함수 분리
